Static build and cross-compilation with Docker #331
