+++ server/common/version_test.go
@@ -2,13 +2,41 @@ package common
 
 import (
 	"fmt"
+	"regexp"
 	"strings"
 	"testing"
 
-	"github.com/root-gg/plik/version"
-
 	"github.com/stretchr/testify/require"
 )
+
+func getVersionRegex() string {
+	return `^\d+\.\d+((\.\d+)?|(\-RC\d+))$`
+}
+
+func validateVersion(t *testing.T, version string, ok bool) {
+	matched, err := regexp.Match(getVersionRegex(), []byte(version))
+	require.NoError(t, err, "invalid version regex")
+
+	if ok {
+		require.True(t, matched, "invalid version regex match")
+	} else {
+		require.False(t, matched, "invalid version regex match")
+	}
+}
+
+func TestValidateVersionRegex(t *testing.T) {
+	validateVersion(t, "1.1", true)
+	validateVersion(t, "1.1.1", true)
+	validateVersion(t, "1.1-RC1", true)
+	validateVersion(t, "1.1.1-RC1", false)
+	validateVersion(t, "1.1-rc1", false)
+}
+
+func TestGetVersion(t *testing.T) {
+	version := GetVersion()
+	require.NotZero(t, version, "missing version")
+	validateVersion(t, version, true)
+}
 
 func TestGetBuildInfo(t *testing.T) {
 	buildInfo := GetBuildInfo()
@@ -28,5 +56,5 @@ func TestGetBuildInfoStringSanitize(t *testing.T) {
 	buildInfo := GetBuildInfo()
 	buildInfo.Sanitize()
 	v := buildInfo.String()
-	require.Equal(t, fmt.Sprintf("v%s", version.Get()), v, "invalid build string")
+	require.Equal(t, fmt.Sprintf("v%s", GetVersion()), v, "invalid build string")
 }
