--- version/version.go
@@ -1,8 +1,0 @@
-package version
-
-const version = "1.3"
-
-// Get return the current package version
-func Get() string {
-	return version
-}
